+++ src/components/Rule/AlterRule.js
@@ -1,11 +1,14 @@
 import React, { PureComponent } from 'react';
+import { connect } from 'react-redux';
+import { compose } from 'recompose';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { toPairs, has, includes } from 'lodash';
+import { toPairs, has, includes, keys, flow, map, fromPairs } from 'lodash';
 import { SortableElement } from 'react-sortable-hoc';
 import DragHandle from './DragHandle';
 import DropDown from './DropDown';
 import Input from './Input';
+import { getVariableRegistry } from '../../selectors/protocol';
 
 const operators = toPairs({
   EXACTLY: 'is Exactly',
@@ -122,6 +125,27 @@ class AlterRule extends PureComponent {
   }
 }
 
+function mapStateToProps(state) {
+  const variableRegistry = getVariableRegistry(state);
+
+  const nodeAttributes = flow(
+    toPairs,
+    nodeTypes => map(
+      nodeTypes,
+      ([nodeType, options]) => [nodeType, keys(options.variables)],
+    ),
+    fromPairs,
+  );
+
+  return {
+    nodeTypes: keys(variableRegistry.node),
+    nodeAttributes: nodeAttributes(variableRegistry.node),
+  };
+}
+
 export { AlterRule };
 
-export default SortableElement(AlterRule);
+export default compose(
+  SortableElement,
+  connect(mapStateToProps),
+)(AlterRule);
