+++ src/components/Rule/EdgeRule.js
@@ -1,11 +1,14 @@
 import React, { PureComponent } from 'react';
+import { connect } from 'react-redux';
+import { compose } from 'recompose';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { toPairs, has, includes } from 'lodash';
+import { toPairs, has, includes, keys, flow, map, fromPairs } from 'lodash';
 import { SortableElement } from 'react-sortable-hoc';
 import DragHandle from './DragHandle';
 import DropDown from './DropDown';
 import Input from './Input';
+import { getVariableRegistry } from '../../selectors/protocol';
 
 const operators = toPairs({
   EXACTLY: 'is Exactly',
@@ -122,4 +125,28 @@ class EdgeRule extends PureComponent {
   }
 }
 
-export default SortableElement(EdgeRule);
+function mapStateToProps(state) {
+  const variableRegistry = getVariableRegistry(state);
+
+  const edgeAttributes = flow(
+    toPairs,
+    edgeTypes => map(
+      edgeTypes,
+      ([edgeType, options]) => [edgeType, keys(options.variables)],
+    ),
+    fromPairs,
+  );
+
+  return {
+    edgeTypes: keys(variableRegistry.edge),
+    edgeAttributes: edgeAttributes(variableRegistry.edge),
+  };
+}
+
+export { EdgeRule };
+
+export default compose(
+  SortableElement,
+  connect(mapStateToProps),
+)(EdgeRule);
+
