+++ src/components/Rule/EgoRule.js
@@ -1,11 +1,14 @@
 import React, { PureComponent } from 'react';
+import { connect } from 'react-redux';
+import { compose } from 'recompose';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { toPairs, includes } from 'lodash';
+import { toPairs, includes, keys } from 'lodash';
 import { SortableElement } from 'react-sortable-hoc';
 import DragHandle from './DragHandle';
 import DropDown from './DropDown';
 import Input from './Input';
+import { getVariableRegistry } from '../../selectors/protocol';
 
 const operators = toPairs({
   EXACTLY: 'is Exactly',
@@ -103,4 +106,17 @@ class EgoRule extends PureComponent {
   }
 }
 
-export default SortableElement(EgoRule);
+function mapStateToProps(state) {
+  const variableRegistry = getVariableRegistry(state);
+
+  return {
+    nodeAttributes: keys(variableRegistry.node.person.variables),
+  };
+}
+
+export { EgoRule };
+
+export default compose(
+  SortableElement,
+  connect(mapStateToProps),
+)(EgoRule);
