--- src/containers/README.md
@@ -1,3 +1,0 @@
-# Containers
-
-Smart container components: https://medium.com/@dan_abramov/smart-and-dumb-components-7ca2f9a7c7d0
