--- src/containers/Rule/index.js
@@ -1,4 +1,0 @@
-export { default as AlterRule } from './AlterRule';
-export { default as EdgeRule } from './EdgeRule';
-export { default as EgoRule } from './EgoRule';
-export { default as NetworkRule } from './NetworkRule';
