--- src/containers/index.js
@@ -1,7 +1,0 @@
-export { default as App } from './App';
-export { default as Protocol } from './Protocol';
-export { default as Start } from './Start';
-export { default as Card } from './Card';
-export { default as ProtocolCard } from './ProtocolCard';
-export { default as EditSkipLogic } from './EditSkipLogic';
-export { default as FilterGroup } from './FilterGroup';
